+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -88,6 +88,7 @@ public class Settings extends Config {
 	public static boolean bypassSpawnBlocking = true;
 	public static boolean deleteUnspawnableShopkeepers = false;
 	public static boolean checkShopInteractionResult = false;
+	public static boolean enableLeftClickInteraction = false;
 
 	public static boolean enableWorldGuardRestrictions = false;
 	public static boolean requireWorldGuardAllowShopFlag = false;