+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/entity/base/BaseEntityShopListener.java
@@ -82,9 +82,15 @@ class BaseEntityShopListener implements Listener {
 	// Key: Player id
 	private final Map<UUID, EntityInteraction> lastEntityInteractions = new HashMap<>();
 
-	BaseEntityShopListener(SKShopkeepersPlugin plugin) {
+	// Tries to detect players left clicking entities.
+	// Only used if left click interactions are enabled in the config, and only on Spigot servers:
+	// For Paper, we have a different implementation in the Paper-specific module.
+	private final PlayerLeftClickEntityListener playerLeftClickEntityListener;
+
+	BaseEntityShopListener(SKShopkeepersPlugin plugin, BaseEntityShops entityShops) {
 		this.plugin = plugin;
 		this.shopkeeperRegistry = plugin.getShopkeeperRegistry();
+		this.playerLeftClickEntityListener = new PlayerLeftClickEntityListener(plugin, entityShops);
 	}
 
 	void onEnable() {
@@ -103,10 +109,14 @@ class BaseEntityShopListener implements Listener {
 				EventPriority.LOWEST,
 				plugin
 		);
+
+		playerLeftClickEntityListener.onEnable();
 	}
 
 	void onDisable() {
 		HandlerList.unregisterAll(this);
+
+		playerLeftClickEntityListener.onDisable();
 	}
 
 	// We want to bypass other plugins by default, so that shops can also be opened in protected
@@ -203,12 +213,22 @@ class BaseEntityShopListener implements Listener {
 			return;
 		}
 
+		this.handleShopkeeperInteraction(player, clickedEntity, shopkeeper);
+	}
+
+	private void handleShopkeeperInteraction(
+			Player player,
+			Entity clickedEntity,
+			AbstractShopkeeper shopkeeper
+	) {
 		// The PlayerInteractAtEntityEvent gets sometimes called additionally to the
 		// PlayerInteractEntityEvent. We only want to handle the interaction once.
 		// However, for certain entities, e.g. armor stands, we only receive the
 		// PlayerInteractAtEntityEvent, so we cannot simply ignore this event type.
 		// We remember the last entity interaction for each player and cancel but ignore the event
 		// if we already handled an interaction with the same entity recently.
+		// This also covers left click interactions, which we might receive in addition to a normal
+		// interaction.
 		var lastEntityInteraction = lastEntityInteractions.computeIfAbsent(
 				player.getUniqueId(),
 				playerId -> new EntityInteraction()
@@ -234,6 +254,25 @@ class BaseEntityShopListener implements Listener {
 	@EventHandler(priority = EventPriority.LOWEST, ignoreCancelled = false)
 	void onEntityInteractAt(PlayerInteractAtEntityEvent event) {
 		this.onEntityInteract(event);
+	}
+
+	// Only called if "enable-left-click-interaction" is enabled:
+	boolean onPlayerLeftClickEntity(Player player, Entity entity) {
+		// If the clicked entity is a complex entity part, we continue with its parent:
+		Entity clickedEntity = EntityUtils.resolveComplexEntity(entity);
+
+		Log.debug(() -> "Player " + player.getName() + " is left clicking "
+				+ clickedEntity.getType() + " at " + clickedEntity.getLocation());
+
+		// Also checks for Citizens NPC shopkeepers:
+		AbstractShopkeeper shopkeeper = shopkeeperRegistry.getShopkeeperByEntity(clickedEntity);
+		if (shopkeeper == null) {
+			Log.debug("  Non-shopkeeper");
+			return false;
+		}
+
+		this.handleShopkeeperInteraction(player, clickedEntity, shopkeeper);
+		return true;
 	}
 
 	@EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)